requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-multipart
numpy
scikit-learn
pillow
python-dotenv
tensorflow
Cython
contextlib2
lxml
matplotlib
pandas
tf_slim
opencv-python
protobuf
